Refactor out einsum graph representations from graph_ops
