+++ autohoot/einsum_graph/README.md
@@ -1,0 +1,10 @@
+This folder includes the graph data structures of einsum (einsum graph), and the transfer between
+einsum expressions and the einsum graph.
+
+1. *union_find*: The union find data structure.
+
+2. *graph_structure*: Contains the data structures used to represent the einsum graph.
+
+3. *graph_generator*: Contains functions to generate einsum graphs from einsum expressions.
+
+4. *expr_generator*: Contains functions to generate einsum expressions from einsum graphs.
