+++ autohoot/einsum_graph/__init__.py
@@ -11,3 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from . import graph_generator
+from . import graph_structure
+from . import expr_generator
+from . import union_find
