+++ autohoot/graph_ops/__init__.py
@@ -14,8 +14,8 @@
 
 from . import graph_als_optimizer
 from . import graph_dedup
-from . import graph_generator
 from . import graph_inv_optimizer
-from . import graph_optimizer
+from . import graph_pruning
 from . import graph_transformer
-from . import union_find
+from . import graph_utils
+from . import optimal_tree
