+++ autohoot/graph_ops/graph_als_optimizer.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 """
     This file will contains the graph transformations and optimizations for 
     alternating least squares.
@@ -20,9 +19,9 @@ import logging
 from autohoot import autodiff as ad
 
 from autohoot.utils import get_all_inputs, find_topo_sort, OutputInjectedMode
-from autohoot.graph_ops.graph_generator import split_einsum, get_common_ancestor, generate_optimal_tree, generate_optimal_tree_w_constraint
+from autohoot.graph_ops.optimal_tree import split_einsum, get_common_ancestor, generate_optimal_tree, generate_optimal_tree_w_constraint
 from autohoot.graph_ops.graph_dedup import dedup, remove_transposes
-from autohoot.graph_ops.graph_transformer import rewrite_einsum_expr
+from autohoot.einsum_graph.expr_generator import rewrite_einsum_expr
 from numpy.core.einsumfunc import _parse_einsum_input
 
 FORMAT = '[%(asctime)-15s %(filename)s:%(lineno)s] %(message)s'
