+++ autohoot/graph_ops/graph_dedup.py
@@ -13,12 +13,13 @@
 # limitations under the License.
 
 from autohoot import autodiff as ad
-import copy, logging
+import logging
 import numpy as np
 import itertools
 
 from autohoot.utils import find_topo_sort, OutputInjectedModeP, replace_node, PseudoNode
 from autohoot.utils import find_topo_sort_p
+from autohoot.einsum_graph.graph_generator import get_disjoint_set
 from numpy.core.einsumfunc import _parse_einsum_input
 from collections import defaultdict
 
@@ -73,56 +74,6 @@ def declone(o_node):
         new_inputs.append(i_node)
     o_node.set_inputs(new_inputs)
     return o_node
-
-
-def get_disjoint_set(node):
-    """
-    Get the disjoint set information of the input einsum node.
-
-    The returned set element has the following [key]:[value] structure:
-    [connected_dims]:[output index].
-
-    connected_dims: tuple of input node dims info connected by one char.
-        Each element in the tuple is a DimInfo object.
-
-    When the list of dims is connected by a contraction char, the output index will be -1.
-    """
-    from autohoot.graph_ops.graph_transformer import generate_einsum_info
-
-    def sort_hash(dim_info):
-        return dim_info.name
-
-    node_copy = copy.deepcopy(node)
-    # this is used to normalize the output node name
-    temp_out_name = "_temp_einsum"
-    node_copy.name = temp_out_name
-
-    uf, _, _ = generate_einsum_info(node_copy)
-    # each set contains the dimension names connected by one char
-    dset = uf.disjoint_set()
-    dset_ret = {}
-
-    for connected_dims in dset:
-        if not any(temp_out_name == dim_info.node_name
-                   for dim_info in connected_dims):
-            # contracted char
-            dset_ret[tuple(sorted(connected_dims, key=sort_hash))] = -1
-        else:
-            output_dim_info, = filter(
-                lambda dim_info: dim_info.node_name == temp_out_name,
-                connected_dims)
-            # uncontracted char
-            connected_dims_ret = tuple(
-                sorted(list(
-                    filter(
-                        lambda dim_info: dim_info.node_name != temp_out_name,
-                        connected_dims)),
-                       key=sort_hash))
-            dset_ret[connected_dims_ret] = output_dim_info.dim_index
-
-    # Note: the value list of dset_ret will have an order of 0, ..., len(node.shape), -1.
-    # It is determined because of the routine in generate_einsum_info.
-    return dset_ret
 
 
 def collapse_symmetric_expr(A, B):
