+++ autohoot/utils.py
@@ -211,7 +211,7 @@ class StandardEinsumExprMode:
         self.node = node
 
     def __enter__(self):
-        from autohoot.graph_ops.graph_transformer import generate_einsum_info
+        from autohoot.einsum_graph.graph_generator import generate_einsum_info
         uf, p_outnode, p_innodes = generate_einsum_info(self.node)
         self.node.uf = uf
         self.p_outnode = p_outnode
@@ -236,28 +236,6 @@ class PseudoNode(object):
     def generate_subscript(self, uf):
         self.subscript = "".join(
             [uf.rootval(dim_info) for dim_info in self.dims_info])
-
-
-class DimInfo(object):
-    """The einsum node dimension information"""
-    def __init__(self, node, dim_index, node_index=None, temp_node_name=None):
-        """
-        dim_index: the index of the dimension in the node.
-        node_index: the index of the node in a specific node list (can be None if not specified).
-        temp_node_name: the temporary name used to override the node name (can be None if not specified).
-        """
-        self.node = node
-        self.dim_index = dim_index
-        self.node_index = node_index
-        if temp_node_name != None:
-            self.node_name = temp_node_name
-        else:
-            self.node_name = node.name
-
-        self.name = f"{self.node_name}-{self.node_index}-{self.dim_index}"
-
-    def __str__(self):
-        return self.name
 
 
 def get_root(nodes):
