+++ examples/mps.py
@@ -20,7 +20,7 @@ from autohoot import backend as T
 
 from autohoot.utils import CharacterGetter
 from tensors.quimb_tensors import rand_mps, ham_heis_mpo, load_quimb_tensors, gauge_transform_mps
-from autohoot.graph_ops.graph_generator import split_einsum
+from autohoot.graph_ops.optimal_tree import split_einsum
 from autohoot.graph_ops.graph_transformer import simplify
 from autohoot.graph_ops.graph_als_optimizer import generate_sequential_optimal_tree
 from numpy.core.einsumfunc import _parse_einsum_input
